requirements.txt
pandas
pandas-datareader
matplotlibplotly
